+++ R/ggwordcloud-wrapper.R
@@ -24,9 +24,9 @@
 #' @return a ggplot
 #' @examples
 #' set.seed(42)
-#' data("love_words_small")
+#' data("love_words_latin_small")
 #'
-#' ggwordcloud(love_words_latin$word, love_words_latin$speakers)
+#' ggwordcloud(love_words_latin_small$word, love_words_latin_small$speakers)
 #' @export
 ggwordcloud <- function(words, freq, scale = c(4, 0.5), min.freq = 3, max.words = Inf,
                         random.order = TRUE, random.color = FALSE, rot.per = 0.1,
@@ -100,9 +100,9 @@ ggwordcloud <- function(words, freq, scale = c(4, 0.5), min.freq = 3, max.words 
 #' @return a ggplot
 #' @examples
 #' set.seed(42)
-#' data("love_words_small")
+#' data("love_words_latin_small")
 #'
-#' ggwordcloud2(love_words_latin[,c("word", "speakers")])
+#' ggwordcloud2(love_words_latin_small[,c("word", "speakers")])
 #' @export
 ggwordcloud2 <- function(data,
                          size = 1,