+++ data-raw/love_words.R
@@ -9,7 +9,7 @@ love_words <- love_words |>
   arrange(desc(speakers))
 love_words_small <- love_words |>  filter(speakers > 50)
 love_words_latin <- love_words |> filter(!is.na(iconv(word, from="utf-8", to="latin1") == word))
-love_words_latin_small <- love_words_latin |> filter(speakers > 25)
+love_words_latin_small <- love_words_latin |> filter(speakers > 50)
 save(love_words, file = "./data/love_words.rda")
 save(love_words_small, file = "./data/love_words_small.rda")
 save(love_words_latin, file = "./data/love_words_latin.rda")